app/requirements.txt
gunicorn==20.1.0
uvicorn[standard]==0.17.6
fastapi==0.75.2
httpx==0.22.0
dnspython==2.2.1
pyyaml==6.0
mime==0.1.0
